Make logs go to stdout; closes #1
Add variables for versions of software installed to make it easier to
see and update.

PureFTPd logs to syslog only, so use
[syslog-stdout](https://github.com/timonier/syslog-stdout) to make
syslog messages go to stdout.  This has the additional benefit that it
handles anything that can use syslog instead of just PureFTPd.

Temporarily install curl and use it to fetch syslog-stdout, instead of
using built-in (busybox) wget as that requires an additional
[ssl_helper](https://busybox.net/downloads/binaries/ssl_helper-x86_64)
for HTTPS, but that seems temporarily unavailable at this time (server
down).