Improve docs on volumes and usage examples
Add section on volumes and what directories in the container are
significant and may need to be preserved, this is pretty standard but
saves looking at the Dockerfile to figure out where things are.

Also add a complete example of using virtual users with the
`pureftpd-auto-users` image to maintain accounts automatically based on
config data.