Re-organise how defaults and settings are handled
Previous implementation was inconsistent as settings were saved in a
location only accessible from the main ftpd container.

Settings are now saved in CONFDIR, which must be a volume if settings
are to be shared with co-operating containers.

A new defaults file is added to the base image (so it can be relied to
always be there) loads those settings, falling back to defaults.

This new defaults file is explicitly loaded in any scripts that need
those settings as well as in init of the co-operating auto-users image
(so it can call pure-pw directly).